app/requirements.txt
fastapi
uvicorn
yfinance
polygon-api-client
kafka-python
python-dotenv
